Cleaning, change to final where that is possible  remove unused fuctions and class memembers

+++ src/net/sf/jmuserver/gs/serverPackage/SLoginAuthAnsfer.java
@@ -1,18 +1,20 @@
 package net.sf.jmuserver.gs.serverPackage;
+
 import java.io.IOException;
 
 public class SLoginAuthAnsfer extends ServerBasePacket {
 
 	private byte f = 0; // flafa
 
-	public static final int  PA_InvaltPassword = 0x00;
-	public static final int PA_PassOK    		=0x01;
-	public static final int PA_AccInvalt           =0x02;
-	public static final int PA_AccAlredyConnected  =0x03;
-	public static final int PA_ServerIsFull    	=0x04;
-	public static final int PA_AccBlacked    	=0x05;
-	public static final int PA_NewVersionOfGamereq =0x06;
-	public static final int PA_NoChargeInfo        =0x09;
+	public static final int PA_InvaltPassword = 0x00;
+	public static final int PA_PassOK = 0x01;
+	public static final int PA_AccInvalt = 0x02;
+	public static final int PA_AccAlredyConnected = 0x03;
+	public static final int PA_ServerIsFull = 0x04;
+	public static final int PA_AccBlacked = 0x05;
+	public static final int PA_NewVersionOfGamereq = 0x06;
+	public static final int PA_NoChargeInfo = 0x09;
+
 	public SLoginAuthAnsfer(byte fl) {
 		super();
 		f = fl;
@@ -22,7 +24,7 @@ public class SLoginAuthAnsfer extends ServerBasePacket {
 	@Override
 	public byte[] getContent() throws IOException {
 		mC1Header(0xf1, 0x01, 0x05);
-		_bao.write((byte) f);
+		_bao.write(f);
 		_bao.write(0x00);
 		System.out.println(f);
 		return _bao.toByteArray();
